Almost fucked everything up because I didn't use source control. Use source control you idiots!
Commit contains the following:
Oscillator - generates a wave (sine/square/saw/triangle) at a given frequency and peak amplitude
EnvelopeADSR - a simple implementation of an ADSR envelope generator
FilterResonant - 4-pole IIR filter (lowpass, highpass,  bandpass)

Three scenes show how to use the classes listed above.
SingleOscSynth - demonstrates how to use the oscillator
ADSROscSynth - demonstrates how to use the envelope with the oscillator
FilterADSROscSynth - demonstrates how to use the filter with the oscillator as well as using the EnvelopeADSR with the filter.

I have no idea how correct everything is but things respond and it's fun to mess with.